fixed button for favorites to work appropriately

+++ app/api/favorites_routes.py
@@ -27,24 +27,19 @@ def create_favorite(id):
 @favorites_routes.route('/delete/<int:id>', methods=['DELETE'])
 @login_required
 def delete_favorite(id):
-    print("FAV TO DELETE!!!")
     fave_to_delete = Favorite.query.get(id)
-    db.session.delete(fave_to_delete)
-    db.session.commit()
-    return {"message": "Favorite Deleted!"}
+
+    if (fave_to_delete):
+
+        db.session.delete(fave_to_delete)
+        db.session.commit()
+        return {"message": "Favorite Deleted!"}
+    else:
+        return {'errors': "No favorite to delete"}
 
 
-# TODO - all favorites needs debugging!
 @favorites_routes.route('/')
 def get_all_favorites():
     favorites = Favorite.query.all()
 
-    print("hitting route")
     return [favorite.to_dict() for favorite in favorites]
-
-# @favorites_routes.route('/<int:id>')
-# @login_required
-# def get_one_favorite():
-#     fave_to_get = Favorite.query.get(id)
-
-#     return {"message": "Favorite Deleted!"}
